refactor(explore): extract BKN handlers into explore-bkn.ts
